Big refactor plus navbarleft functionality
