+++ src/Global.css
@@ -7,7 +7,7 @@
     --color-button-red: #990033;
     --color-lite-red: #A72229FF;
 
-    --color-white: #FBEBECFF;
+    --color-white: #FFFFFF;
 
     --color-lightbox: rgba(0, 0, 0, 0.8);
