+++ src/pages/Home/components/FileSearchComponents/FileLine.css
@@ -6,7 +6,7 @@
 
 .file-line {
     display: grid;
-    grid-template-columns: repeat(2, 1fr);
+    grid-template-columns: repeat(3, 1fr);
     border-bottom: 1px solid var(--color-white);
     background-color: transparent;
     color: white;
