+++ src/pages/Home/components/FileSearchComponents/FileSearch.css
@@ -1,37 +1,3 @@
-/* Holds the side-bar and main-bar */
-
-@import url("../../Global.css");
-
-.full-screen {
-    max-height: 100vh;
-    overflow-y: hidden;
-    user-select: none;
-}
-.page-container {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    background-color: var(--color-bg);
-}
-
-.side-bar {
-    display: flex;
-    width: 15%;
-}
-
-.main-bar {
-    display: flex;
-    flex-direction: column;
-    width: 85%;
-    padding: 10px;
-    background-color: var(--color-dark-gray);
-    border-radius: 10px;
-    margin-right: 1rem;
-    overflow-y: auto;
-    height: 86vh;
-}
-
-
 .file-query {
     display: flex;
     flex-direction: column;
@@ -97,29 +63,29 @@
     box-shadow: 0 8px 16px 0 var(--color-lightbox);
     z-index: 1;
     border-radius: 5px;
-  }
+}
 
-  .dropdown-content a {
+.dropdown-content a {
     color: var(--color-gray);
     padding: 12px 16px;
     text-decoration: none;
     display: block;
     text-align: center;
-  }
+}
 
-  .dropdown-content a:hover {
+.dropdown-content a:hover {
     background-color: var(--color-dark-gray);
-  }
+}
 
 
-  .dropdown-button:focus {
+.dropdown-button:focus {
     background-color: var(--color-lite-red);
-  }
+}
 
-  .files-per-page-dropdown {
-      padding: 1rem 1.5rem 1rem 1rem;
-      margin-bottom: -1rem;
-      margin-left: auto;
+.files-per-page-dropdown {
+    padding: 1rem 1.5rem 1rem 1rem;
+    margin-bottom: -1rem;
+    margin-left: auto;
 }
 
 .files-per-page-dropdown label {
@@ -221,4 +187,3 @@
     width: 24px;
     height: 24px;
 }
-
