+++ src/pages/Home/components/FileUpload.css
@@ -18,7 +18,7 @@
     background: var(--color-dark-gray);
     color: var(--color-white);
     padding: 20px;
-    border-radius: 5px;
+    border-radius: 10px;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.3);
     text-align: center;
     max-width: 40vw; /* Optional: limit the width of the lightbox */
