+++ src/pages/Home/components/SavedFilesComponents/SavedFilesPage.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const SavedFilesPage = () => {
+    return (
+        <div>
+            <p>lets go</p>
+        </div>
+    );
+}
+
+export default SavedFilesPage
